Fix tax calculation precision loss and clean up code
- Changed tax fields from u8 to f64 to preserve decimal precision (e.g., 25.7% won't truncate to 25%)
- Removed redundant calculate_taxes_from_state_changes method in tax_detector.rs
- Tax detector now correctly uses centralized tax calculation functions from tax_calculator.rs
- Fixed database module rename (database -> db_writers) in mempool_signal_detector binary
- Cleaned up unused imports in tax_detector.rs

The tax calculation formula is mathematically correct:
- Buy Tax: (1 - tokens_received/tokens_sent_by_pool) × 100
- Sell Tax: (1 - eth_received/eth_sent_by_pool) × 100

+++ reth/mempool_processor/src/signal_detector/signal_manager.rs
@@ -522,8 +522,8 @@ impl SignalManager {
                         pool_address,  // Now includes the specific pool
                         pool_type,     // Pool type (V2, V3, V4)
                         creator_address: trading_signal.executor.clone(),
-                        buy_tax: calculated_buy_tax.unwrap_or(0.0) as u8,
-                        sell_tax: calculated_sell_tax.unwrap_or(0.0) as u8,
+                        buy_tax: calculated_buy_tax.unwrap_or(0.0),
+                        sell_tax: calculated_sell_tax.unwrap_or(0.0),
                         timestamp: chrono::Utc::now().timestamp() as u64,
                     }));
                 }
